Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-elem-14-t-drt-expected.txt

calcMode="discrete"
keyTimes="0;.2;.4;.6"
Time (s):
0
2
4
6
$Revision: 1.6 $

FAIL A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times assert_approx_equals: expected 30 +/- 1 but got 180
